NET-85: moved pinger to separate package to prevent future import cycles, fixed network compilation
