+++ network/hostnetwork/transport_test.go
@@ -29,6 +29,7 @@ import (
 	"github.com/insolar/insolar/network"
 	"github.com/insolar/insolar/network/transport/host"
 	"github.com/insolar/insolar/network/transport/packet/types"
+	"github.com/insolar/insolar/network/utils"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/require"
 )
@@ -185,7 +186,7 @@ func TestNewHostTransport(t *testing.T) {
 		_, err := t1.SendRequest(request, core.NewRefFromBase58(ID2))
 		require.NoError(t, err)
 	}
-	success := network.WaitTimeout(&wg, time.Second)
+	success := utils.WaitTimeout(&wg, time.Second)
 	require.True(t, success)
 }
 
@@ -248,7 +249,7 @@ func TestHostTransport_SendRequestPacket2(t *testing.T) {
 
 	_, err = t1.SendRequest(request, core.NewRefFromBase58(ID2))
 	require.NoError(t, err)
-	success := network.WaitTimeout(&wg, time.Second)
+	success := utils.WaitTimeout(&wg, time.Second)
 	require.True(t, success)
 }
 
@@ -359,7 +360,7 @@ func TestHostTransport_WrongHandler(t *testing.T) {
 	require.NoError(t, err)
 
 	// should timeout because there is no handler set for Ping packet
-	result := network.WaitTimeout(&wg, time.Millisecond*10)
+	result := utils.WaitTimeout(&wg, time.Millisecond*10)
 	require.False(t, result)
 }
